separating customers logic
