+++ modules/customers.c
@@ -24,6 +24,79 @@ customer *createNewCustomer()
 	return newCustomer;
 }
 
+// Copy customer?
+customer *createCustomer(customer *from)
+{
+  customer *newCustomer = malloc(sizeof(customer));
+  newCustomer->code = from->code;
+  strcpy(newCustomer->name, from->name);
+  strcpy(newCustomer->email, from->email);
+  strcpy(newCustomer->phone, from->phone);
+  strcpy(newCustomer->address, from->address);
+  return newCustomer;
+}
+
+customer *searchCustomerById()
+{
+  int *code;
+  int found = 0;
+  customer_node *current = customers->head;
+  printf("Digite el codigo de cliente a buscar: ");
+  scanf("%d", &code);
+  getchar();
+  while (current != NULL)
+  {
+    if (current->value->code == code)
+    {
+      return current->value;
+    }
+    current = current->next;
+  }
+  printf("\nCliente con codigo '%d' no encontrado...\n\n", code);
+  return NULL;
+}
+
+customer *searchCustomerByName()
+{
+  char name[50];
+  customer_node *current = customers->head;
+  customer_list *foundResults = malloc(sizeof(customer_list));
+  printf("Digite el nombre del cliente a buscar: ");
+  readString(name, 50);
+  while (current != NULL)
+  {
+    if (strcmp(current->value->name, name) == 0)
+    {
+      customer *newCustomer = createCustomer(current->value);
+      addCustomer(foundResults, newCustomer);
+    }
+    printf("a %d", &foundResults->size);
+    current = current->next;
+    printf("b");
+  }
+  printf("\nClientes con nombre '%s' encontrados: %d\n", name, foundResults->size);
+  if (foundResults->size == 1)
+  {
+    return foundResults->head->value;
+  }
+  else if (foundResults->size > 1)
+  {
+    printf("----------\n");
+    printf(customers->head->value->name);
+    printf("\n");
+    printf(foundResults->head->value->name);
+    printf("----------\n");
+    printf("\nSe encontraron mas de un cliente con el nombre '%s'\n", name);
+    printf("N de clientes encontrados: %d\n", foundResults->size);
+    // printCustomers(foundResults);
+  }
+  else
+  {
+    printf("\n Cliente con nombre '%s' no encontrado\n\n", name);
+  }
+  return NULL;
+}
+
 customer *searchCustomer()
 {
 	int option;
@@ -108,4 +181,35 @@ customer *getCustomer()
 		}
 	}
 	return NULL;
+}
+
+void printCustomer(customer *customer)
+{
+  printf("\n\----DATOS DE CLIENTE----\n");
+  printf("\Mem add: %d", &customer);
+  printf("\nCodigo %d", customer->code);
+  printf("\nNombre %s", customer->name);
+  printf("\nDireccion %s", customer->address);
+  printf("\nEmail %s", customer->email);
+  printf("\nTelefono %s\n\n", customer->phone);
+}
+
+void printCustomers(customer_list *list)
+{
+  customer_node *current = list->head;
+  if (current != NULL)
+  {
+    printf("---Lista de clientes---\n");
+    while (current != NULL)
+    {
+      printf("cliente: %d\n", current->value->code);
+      printCustomer(current->value);
+      current = current->next;
+    }
+  }
+  else
+  {
+    printf("Lista de clientes vacia...");
+  }
+  system("pause");
 }
